requirements.txt
streamlit==1.52.2
google-generativeai
httpx==0.28.1
python-docx==1.2.0
